cherry_picking/requirements.txt
pulp-rpm-client==3.14.2
pulpcore-client==3.14.5
PyYAML==5.4.1
colorama==0.4.4
